Delete HasNext.
It doesn't make much sense to ask for the length of a stream stored in
an accumulator (e.g., inside StmBuild), since the length changes over
time. Instead, we should explicitly use a counter (this is what would
need to happen in hardware anyway). HasNext was only used in such
invalid situations, so I don't see a use for it anymore.