Remove constant-valued accumulator elems
This comes up especially often when dealing with StmMap().
StmMap() uses counters to know when to repeat f.
If f takes one element in and produces one element (e.g., if it's a
scalar --> scalar function), it always repeats immediately.
Therefore, StmMap() sometimes leaves a bunch of unnecessary accumulator
elements in the final stream.
This was causing some of the stream fusion tests to fail.