Change order of args in StmFold function
Make it like fold in Haskell or List.fold_left in OCaml, with the
function having the signature `b -> a -> b` rather than `a -> b -> b`.
It'll probably be easier to remember if it's consistent with prominent
existing languages.
I think it'll also make the multi-dimensional implementation a bit
easier.