Implement big-step evaluator and use it in tests
I think this is a bit more reliable because the evaluator is simple and
basically defines the semantics of the language, whereas the partial
evaluator is getting increasingly complex as more optimization rules are
added.

The evaluator also appears to be a bit faster than the partial
evaluator, so the tests take several seconds less.