Add test for StmReverse(StmReverse(s))
Unfortunately that doesn't work yet.  It seems
StmInductionVarRemovalPass is unable to find a closed form due to the
fact that, in this case, there are *four* accumulator elements that all
depend on each other.  There are currently no patterns for dealing with
such a large set of mutually-dependent elements and I'd like to find a
way to deal with that which doesn't require adding another special case.