Tweak semantics of reading empty stream
The test case StmMap:2D-2D:StmShiftLeft was failing, seemingly because
it was evaluating something like
```
Tuple(StmNext(<empty>)().__1, False)().__1
```

In practice, I plan to implement the hardware in such a way that, when
it is not ready (including when empty) and when the valid signal is
lowered, the output is zeroed-out. Therefore, maybe it shouldn't be an
error at all to read from an empty stream, you'll just always get zero
(or the default value for the given type).