Make value in IntCst a Long
In some test cases that were using u32 and i33, the value was
overflowing in Scala. With Long, this problem should at least go away as
long as you stay below 64 bits. In theory you could use BigInt to
completely solve the problem, but this is a bit less convenient (e.g.,
pattern matching on BigInt seems a bit unwieldy:
https://stackoverflow.com/questions/21578307/pattern-matching-bigints).