Revert to explicit reshaping
Implicit reshaping seems more convenient at first glance but
unfortunately I don't think it will work out.

Suppose the type of a sum is inferred to be the same as that of its
operands. Suppose `x : i8`, `y : u3`, and we have the expression `7:u3 +
x + y + -1*x`. The type checker will infer the type `i8` for this sum
and if we were to set `x = 0:i8` and `y = 1:u3`, then we would get
`8:i8`.  But the arithmetic simplifier may simplify this to `7:u3 + y`.
The type checker would infer the type `u3` to this expression, and then
plugging in `y = 1:u3` would result in overflow!

Suppose instead that each sum itself was annotated with a type, so that
the inputs can be whatever type you want while still doing the addition
with enough bits. But that information gets lost after converting to
`ArithExpr` for simplification and I don't see any way to recover it.