Delay exceptions in evaluator
Rather than immediately throwing an exception if the evaluator
encounters undefined behaviour (e.g., out-of-bounds vector access), keep
going but record the problem. If the problem occurs in the branch of a
MUX that is not used, then there's no problem. But if the undefined
value does seem to be used, throw an exception at the very end.