Write general helper method for annotating function literals
For example, consider
```
StmMap(s: stm[u8, n], (x: ???) => f(x))
```
Even though the function input type annotation is missing, it is trivial
to infer that it should be u8 in this case. This sort of thing is
common, especially with the Aetherling benchmarks, so it seems helpful
to have one helper method that adds the annotation to the function if
it's missing.