Simplify input streams in StreamTests
In some stream tests, I need some input stream but I don't care very
much what it is. In such cases I would often use something like
```
StmMap(StmCount(n), i => StmMap(StmCount(m), j => ...))
```
But this construction turns out to be pretty slow and if it happens to
be buggy then a lot of tests will fail. Maybe it's better to just pass
in a `StmLiteral` in such cases.