Make streamifier a bit less stupid
If there are input which were originally data rather than streams, the
streamifier must update the StmBuilds to take one cycle to read that
data and store it in an accumulator. Now it will only add this extra
logic in StmBuilds that actually refer to that data input, rather than
just blindly adding it everywhere.