Investigate slow optimization for smallconv2d
The partial evaluation step at the beginning of the simplification
process seems to account for a large part of that slowdown. At first
glance, it seems like there's a lot of repeated partial evaluation of
the same expression, so maybe caching would help. Unfortunately, just
adding a WeakHashMap in the partial evaluator slows things down for some
reason. It may also introduce type errors, since the `equals` method for
`Expr` currently ignores types.