Try to speed up equals() and hashCode() in StmBuild
StmBuild.hashCode has to do a fair bit of work so that it is insensitive
to the naming of accumulator variables. Making the hashCode a lazy val
rather than a method seems to speed things up a bit. This should also
make equals() faster, since equals() uses hashCode() as a quick test to
discard obviously non-equal objects.

The same goes for Function and LetStm, which also involve variable
binding.