Ensure VHDL design always buffers output
There was an edge case (a LetStm where the bound variable is unused and
the output is just the input) where the VHDL design just passed along
the input unchanged. However, the testbench assumes the output will be
stable across the whole clock cycle (so it can check the output at the
falling edge, in the middle of the cycle).