Manually patch SHIR top-level component
Rather than subtracting the resource usage of the hard-coded inputs from
the total resource usage, just manually update the top-level component
so that ports are exposed for the inputs.

This will yield more accurate measurements of the resource usage,
reduce synthesis time, and may allow me to write testbenches to also
automatically measure the latency.